+++ oneflow/core/framework/op_expr_helper.h
@@ -97,8 +97,8 @@ Maybe<one::UserOpExpr> CastOp(const DataType& to_type, const std::string& name);
 Maybe<one::UserOpExpr> CastLikeOp();
 Maybe<one::UserOpExpr> CastLikeOp(const std::string& name);
 
-Maybe<one::UserOpExpr> CopyOp(const std::string& device_type, const int device_id);
-Maybe<one::UserOpExpr> CopyOp(const std::string& device_type, const int device_id,
+Maybe<one::UserOpExpr> CopyOp(const std::string& device_type, const int64_t device_id);
+Maybe<one::UserOpExpr> CopyOp(const std::string& device_type, const int64_t device_id,
                               const std::string& name);
 
 Maybe<one::UserOpExpr> NormalizationGradOp(const int32_t& axis, const float& epsilon);
